Complete the Beer webhook receiver
NOTE: It's possible that in the event of a new beer, it was added at
around the same time as its parent brewery (i.e. both the beer and its
brewery are new). In this case, we might receive the beer webhook before
the parent brewery's webhook. I've posted on the mailing list about this
possibility to make sure that it is one. If so, this webhook receiver
should probably be edited to create a Brewery webhook receiver for
adding breweries.

Signed-off-by: David Celis <me@davidcel.is>